+++ HISTORY.md
@@ -60,6 +60,8 @@
 ### 2022-07-11
 * extensions: fixed #3065
 * threads: added support to the OF reporter for binding a ring to another one, e.g. THIS SCRIPT to access its local variables
+* byob: embed blocks code into all custom block definition pics
+
 ### 2022-07-04
 * blocks, gui: directly import embedded blocks from a smart pic if the pic is dragged and dropped onto a scripting area or palette - otherwise import the pic as costume (with embedded blocks)
 * gui: import smart pic as costume via "Import..." item in the project menu